ConsultaVehiculosAPI/requirements.txt
asgiref==3.8.1
Django==5.0.8
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
